refactor: split mini-ship-item and dedupe styling with ship-parts
Splits the 606-line mini-ship-item.tsx into sibling modules:
styled-components (all mini styled defs) / mini-slotitems (tooltip
equipment list) / mini-ship-row / mini-squard-row, with mini-ship-item.tsx
kept as a barrel re-exporting MiniShipRow, MiniSquardRow and ShipAvatar so
the plugin-facing surface is unchanged. All semantic classNames
(ship-item, ship-name, hp-progress, ...) are preserved.

Deduplication against ship-parts:
- Gradient: the mini copy (identical css except grid-row, down to the
  copy-pasted comment) becomes MiniGradient = styled(Gradient) with only
  the grid-row override.
- New shared overAvatarText() css helper in ship-parts/styled-components
  expresses the white-over-avatar text block; the mini panel uses it with
  its tight 3px shadow. (Adopting it in the five equivalent blocks of the
  main ship panel is left as a follow-up.)
- The other name collisions with ship-parts (ShipItem, ShipAvatar,
  ShipName, ShipHP, ShipCond, StatusLabelContainer) are genuinely
  different css, so the mini versions are renamed Mini* instead of merged.
- slotitemsDataSelectorFactory is now exported from ship/slotitems and
  shared: the mini copy was a strict subset (same inputs, minus the
  exslotUnlocked field, which mini ignores).
- miniShipRowDataSelectorFactory intentionally stays separate from
  shipRowDataSelectorFactory: sharing would subscribe every mini row to
  state.const and re-render them all on master-data refresh. A comment
  records the relationship.
